+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Remove `elliptic` dependency, decompressing a public key can be done using `crypto`
 * Add support for requesting & receiving new blocks
+* The first transaction in a block indicates who mined it
 
 ## 0.1.0 (2018-02-24)